Use customizer callback to further modify the dependency metadata
The default metadata from Spring Boot Initializr is in
config/application.yml (it could go in other places, but that location
means it is picked up by default). The cloud dependencies are added
from the local application.yml (in cloud.dependencies.*).

The Cloud versions from cloud.versions are used in the "bootVersion" of
the vanilla Initializr service. This means that the versionRanges (if any
in the default dependency metadata are wrong because they match Boot versions.
We adjust for that in the customizer as well, matching on a cloud version
from cloud.bootVersions.

Fixes gh-3